Move secondary trainer extraction into SecondaryTrainer model
