+++ evasys-eai/src/test/java/de/muenchen/evasys/client/EvaSysClientTest.java
@@ -2,7 +2,6 @@ package de.muenchen.evasys.client;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
-import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyInt;
@@ -14,11 +13,9 @@ import static org.mockito.Mockito.when;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.sap.document.sap.rfc.functions.ZLSOSTEVASYSRFC;
-import de.muenchen.evasys.exception.EvaSysException;
 import de.muenchen.evasys.mapper.SapEvaSysMapper;
 import de.muenchen.evasys.model.SecondaryTrainer;
 import jakarta.xml.ws.Holder;
-import java.util.List;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -50,19 +47,6 @@ public class EvaSysClientTest {
     @BeforeEach
     public void setup() {
         evaSysClient = new EvaSysClient(soapPortMock, mapper);
-    }
-
-    private ZLSOSTEVASYSRFC createData(
-            String id, String anrede, String titel,
-            String vorname, String nachname, String email) {
-        ZLSOSTEVASYSRFC data = new ZLSOSTEVASYSRFC();
-        data.setSEKTRAINERID(id);
-        data.setSEKTRAINERANREDE(anrede);
-        data.setSEKTRAINERTITEL(titel);
-        data.setSEKTRAINERVNAME(vorname);
-        data.setSEKTRAINERNNAME(nachname);
-        data.setSEKTRAINERMAIL(email);
-        return data;
     }
 
     @Test
@@ -247,88 +231,6 @@ public class EvaSysClientTest {
         assertEquals("max@example.com", captured.getMSEmail());
         assertEquals(1, captured.getMNFbid());
         assertEquals(1, captured.getMNAddressId());
-    }
-
-    @Test
-    void testReturnsEmptyListWhenNoSecondaryTrainer() {
-        ZLSOSTEVASYSRFC trainingData = new ZLSOSTEVASYSRFC();
-        trainingData.setSEKTRAINERID(null);
-
-        List<SecondaryTrainer> trainers = evaSysClient.extractSecondaryTrainers(trainingData);
-
-        assertTrue(trainers.isEmpty());
-    }
-
-    @Test
-    void testParsesSingleSecondaryTrainer() {
-        ZLSOSTEVASYSRFC trainingData = createData(
-                "2",
-                "2",
-                "Prof.",
-                "Erika",
-                "Musterfrau",
-                "erika@example.com");
-
-        List<SecondaryTrainer> result = evaSysClient.extractSecondaryTrainers(trainingData);
-
-        assertEquals(1, result.size());
-        SecondaryTrainer t = result.get(0);
-
-        assertEquals("2", t.id());
-        assertEquals("2", t.anrede());
-        assertEquals("Prof.", t.titel());
-        assertEquals("Erika", t.vorname());
-        assertEquals("Musterfrau", t.nachname());
-        assertEquals("erika@example.com", t.email());
-    }
-
-    @Test
-    void testParsesMultipleSecondaryTrainers() {
-        ZLSOSTEVASYSRFC trainingData = createData(
-                "11; 22",
-                "1; 2",
-                "Dr.; Prof.",
-                "Max; Erika",
-                "Mustermann; Musterfrau",
-                "max@example.com; erika@example.com");
-
-        List<SecondaryTrainer> result = evaSysClient.extractSecondaryTrainers(trainingData);
-
-        assertEquals(2, result.size());
-
-        SecondaryTrainer t1 = result.get(0);
-        SecondaryTrainer t2 = result.get(1);
-
-        assertEquals("11", t1.id());
-        assertEquals("22", t2.id());
-
-        assertEquals("1", t1.anrede());
-        assertEquals("2", t2.anrede());
-
-        assertEquals("Dr.", t1.titel());
-        assertEquals("Prof.", t2.titel());
-
-        assertEquals("Max", t1.vorname());
-        assertEquals("Erika", t2.vorname());
-
-        assertEquals("Mustermann", t1.nachname());
-        assertEquals("Musterfrau", t2.nachname());
-
-        assertEquals("max@example.com", t1.email());
-        assertEquals("erika@example.com", t2.email());
-    }
-
-    @Test
-    void testThrowsExceptionWhenListLengthsAreInconsistent() {
-        ZLSOSTEVASYSRFC trainingData = createData(
-                "11; 22",
-                "1", // inconsistent → only 1 value
-                "Dr.; Prof.",
-                "Max; Erika",
-                "Mustermann; Musterfrau",
-                "max@example.com; erika@example.com");
-
-        assertThrows(EvaSysException.class, () -> evaSysClient.extractSecondaryTrainers(trainingData));
     }
 
     @Test
